- Creazione mfs Product; - Creazione relazione tra Product e Location (N:N); - Aggiunte (ma per il momento nascoste) voci per Prodotti e Articoli; - Aggiunto numero di prodotti all'interno di una Location nella tabella in locations.index; - Aggiunta la lista dei prodotti presenti in una location in locations.show; - Predisposto pulsante per cambiare la location ad un prodotto in locations.show (non ancora funzionante);

+++ app/Http/Livewire/Pages/Locations/Index.php
@@ -22,14 +22,14 @@
 		}
 
 		public function delete(Location $location) {
-//			if($location->products()->count()) {
-//				$this->dispatchBrowserEvent('open-notification', [
-//					'title'    => __('Errore'),
-//					'subtitle' => __('L\'ubicazione non può essere cancellata perché contiene dei prodotti'),
-//					'type'     => 'error'
-//				]);
-//				return false;
-//			}
+			if($location->products()->count()) {
+				$this->dispatchBrowserEvent('open-notification', [
+					'title'    => __('Errore'),
+					'subtitle' => __('L\'ubicazione non può essere cancellata perché contiene dei prodotti'),
+					'type'     => 'error'
+				]);
+				return false;
+			}
 			$location->delete();
 			$this->emitSelf('$refresh');
 			$this->dispatchBrowserEvent('open-notification', [
